+++ src/index.ts
@@ -4,4 +4,5 @@ export * from './components/SubMenu';
 export * from './components/MenuItem';
 export * from './components/ProSidebarProvider';
 export * from './hooks/useProSidebar';
+export * from './utils/utilityClasses';
 export type { CSSObject } from 'styled-components';